+++ ContentPatcher/Framework/Migrations/Migration_1_18.cs
@@ -1,6 +1,8 @@
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
+using ContentPatcher.Framework.Conditions;
 using ContentPatcher.Framework.ConfigModels;
+using Pathoschild.Stardew.Common.Utilities;
 using StardewModdingAPI;
 
 namespace ContentPatcher.Framework.Migrations
@@ -14,7 +16,13 @@ namespace ContentPatcher.Framework.Migrations
         *********/
         /// <summary>Construct an instance.</summary>
         public Migration_1_18()
-            : base(new SemanticVersion(1, 18, 0)) { }
+            : base(new SemanticVersion(1, 18, 0))
+        {
+            this.AddedTokens = new InvariantHashSet
+            {
+                ConditionType.I18n.ToString()
+            };
+        }
 
         /// <summary>Migrate a content pack.</summary>
         /// <param name="content">The content pack data to migrate.</param>